Apply review suggestions, and other smaller fixes.
