+++ helix-core/src/completion.rs
@@ -1,9 +1,12 @@
+use std::borrow::Cow;
+
 use crate::Transaction;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct CompletionItem {
     pub transaction: Transaction,
-    pub label: String,
+    pub label: Cow<'static, str>,
+    pub kind: Cow<'static, str>,
     /// Containing Markdown
     pub documentation: String,
 }
