+++ helix-view/src/document.rs
@@ -1713,7 +1713,7 @@ impl Document {
         self.version
     }
 
-    pub fn supports_path_completion(&self) -> bool {
+    pub fn path_completion_enabled(&self) -> bool {
         self.language_config()
             .and_then(|lang_config| lang_config.path_completion)
             .unwrap_or_else(|| self.config.load().path_completion)
